API: Controller complated

+++ OnlineStoreManagementSystem.Data/IRepositories/IGenericRepository.cs
@@ -11,7 +11,7 @@ namespace OnlineStoreManagementSystem.Data.IRepositories
         IQueryable<T> GetAll(Expression<Func<T, bool>> expression, string[] includes = null, bool isTracking = true);
         ValueTask<T> GetAsync(Expression<Func<T, bool>> expression, bool isTracking = true, string[] includes = null);
         ValueTask<T> CreateAsync(T entity);
-        ValueTask<bool> DeleteAsync(int id);
+        ValueTask<bool> DeleteAsync(long id);
         T Update(T entity);
         ValueTask SaveChangesAsync();
     }
