+++ OnlineStoreManagementSystem.Service/Interfaces/Attachments/IAttachmentService.cs
@@ -1,10 +1,6 @@
 ﻿using OnlineStoreManagementSystem.Domain.Entitties.Attachments;
 using OnlineStoreManagementSystem.Service.DTOs.Attachments;
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace OnlineStoreManagementSystem.Service.Interfaces.Attachments
@@ -12,7 +8,7 @@ namespace OnlineStoreManagementSystem.Service.Interfaces.Attachments
     public interface IAttachmentService
     {
         ValueTask<Attachment> UploadAsync(AttachmentForCreationDTO dto);
-        ValueTask<Attachment> UpdateAsync(int id, Stream stream);
+        ValueTask<Attachment> UpdateAsync(long id, Stream stream);
         ValueTask<Attachment> CreateAsync(string filePath);
     }
 }
