+++ OnlineStoreManagementSystem.Service/Interfaces/Products/IProductService.cs
@@ -1,12 +1,8 @@
 ﻿using OnlineStoreManagementSystem.Domain.Entitties.Products;
-using OnlineStoreManagementSystem.Domain.Entitties.Users;
 using OnlineStoreManagementSystem.Service.DTOs.Products;
-using OnlineStoreManagementSystem.Service.DTOs.Users;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace OnlineStoreManagementSystem.Service.Interfaces.Products
@@ -16,7 +12,7 @@ namespace OnlineStoreManagementSystem.Service.Interfaces.Products
         ValueTask<IEnumerable<Product>> GetAllAsync(Expression<Func<Product, bool>> expression = null);
         ValueTask<Product> GetAsync(Expression<Func<Product, bool>> expression);
         ValueTask<Product> CreateAsync(ProductForCreationDTO dto);
-        ValueTask<bool> DeleteAsync(int id);
-        ValueTask<Product> UpdateAsync(int id, ProductForCreationDTO dto);
+        ValueTask<bool> DeleteAsync(long id);
+        ValueTask<Product> UpdateAsync(long id, ProductForCreationDTO dto);
     }
 }
