+++ OnlineStoreManagementSystem.Service/Interfaces/Users/IUserProductService.cs
@@ -2,9 +2,7 @@
 using OnlineStoreManagementSystem.Service.DTOs.Users;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace OnlineStoreManagementSystem.Service.Interfaces.Users
@@ -14,6 +12,6 @@ namespace OnlineStoreManagementSystem.Service.Interfaces.Users
         ValueTask<IEnumerable<UserProduct>> GetAllAsync(Expression<Func<UserProduct, bool>> expression = null);
         ValueTask<UserProduct> GetAsync(Expression<Func<UserProduct, bool>> expression);
         ValueTask<UserProduct> CreateAsync(UserProductForCreationDTO dto);
-        ValueTask<bool> DeleteAsync(int id);
+        ValueTask<bool> DeleteAsync(long id);
     }
 }
